LINK-Bot/pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
pyowm = "^3.3.0"
yahoo-fin = "^0.8.9"
discord = "^1.7.3"
pytarkov = "^1.1.2"
google-cloud-vision = "^2.7.2"
uberduckapi = "^0.1.18"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"